[ot] python/qemu: ot.util.misc: fix Git subprocess error handling
subprocess.check_output may raise a SubprocessError, which is not a subclass of OSError.

For example, when a git repository exists but no tag has been defined, git "describe"
command exits with an error, which is translated into a (subclass of) SubprocessError
exception by the subprocess module.

Handle any SubprocessError exception.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>