[ot] hw/opentitan: ot_i2c: use matched address for acq, not bus address
This is a bugfix for target mode address masking.

Previously, the OT I2C's bus address would be used for the address byte
on a start condition. Now that address masking is implemented, the
address that matched may not be exactly the bus address (as some bits
are ignored). Hardware presents the matched address in the start
condition byte so that software can perform different actions based on
different target transaction addresses.

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>