[ot] hw/core: loader.c: make generic loader image loadable at will
Image file loaded with "-device loader,file=..." are copied back to their
destination device when the generic QEMU reset feature is invoked.

As OpenTitan device use a specialized reset scheme, we may need to copy the so-called
"ROM" images (as QEMU non-modifiable image, not related to OT ROM controller images)
at will.

Add a public API to trigger this image-to-device copy since the default API is private.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>