[ot] hw/opentitan: ot_flash: Use IO region for host reads by default
Since Earlgrey's flash is execute-in-place (XIP) and frequently mutated
via bootstrapping, it is important that the default is to use an MMIO
region for host reads from flash memory, as otherwise TB caching in the
TCG could cause invalid host instruction execution and mutating code
(common when e.g. bootstrapping over SPI).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>