[ot] hw/opentitan: ot_pwrmgr: implement initial support for clock domain crossing
Several register properties are not used as-is, as they drive features implemented in
the slow clock domain, while the registers themselves are in the default, "fast" clock
domain.

It is required for the guest SW to write and poll the CFG_CDC_SYNC register to
commit new settings to the slow clock domain.

QEMU was not supporting this feature, the guest SW could then miss to synchronize
these features. Application would have run fine on QEMU but not on the real HW.

Note that the doc is out-of-date, and only the RTL can provide a list of the actual
slow clock domain features gated by the CFG_CDC_DOMAIN.

Also update the guest error message to give a hint when reset cannot be successfully
enabled in this case.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>