[ot] .github/workflows: ci: Add Earlgrey ROM_EXT CI regression
With upstream support, there are now a variety of ROM_EXT end-to-end
tests that we might expect to pass or at least be flaky. To ensure that
regressions on these more complicated tests are being caught, add a
regression run for the non-SiVal ROM_EXT environment, which most of
these tests use.

Since this is a separate job in the regression workflow it will be run
in parallel by a separate runner, and so should not increase existing CI
times (unless the time to execute all ROM_EXT test overshadows the time
for executing other existing execution environments).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>