[ot] tests/opentitan: earlgrey-tests: Add new passing ROM_EXT tests
Add newly passing ROM_EXT tests to be checked in Earlgrey regressions
now that these are properly supported in upstream OpenTitan. Uses the
increased timeout to ensure that even when running in parallel on slower
runners, these tests should pass most of the time.

Mark the couple of tests that are most likely to be flaky due to strict
timing checks in the ROM as such, to hopefully avoid CI flakiness.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>