[ot] tests/opentitan: earlgrey-tests: Mark sigverify test flaky
We expect this test to pass, but recent upstream changes increase the
time to emulate this test by a lot and it will often take 2-5+ minutes
to execute depending on host performance and load.

Mark it flaky instead of removing it because it could theoretically pass
if the runner is fast enough, but it is not expected to pass with the
standard 60 second timeout we give.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>