[ot] hw/opentitan: ot_entropy_src: add support for multiple versions
Support either version 2 or 3. Both versions are mostly similar, however a "REV" register
has been added in v2 then removed in v3. This register stands at the 8th register slot.
This causes most of the register map to shift down on v3 compared to v2. The version
upgrade also introduces a couple of bitfield permutations and width changes in other
registers.

Add a new property to select the version to use.

To address the register map shifting, split the register map in 3 segments:
* lo: for registers before the REV register, whose base address is the same for both
  versions
* rev: for the register which has been introduced in v2 and removed in v3
* hi: for the registers located after the REV register, whose base address depends
  on the presence of the REV register, i.e. whether emulutating v2 or v3
Use QEMU sub memory regions to map each register segment to the proper location based
on the selected version. The addresses of each segment defined with REG32 macros are
made relative to the start location of the matching segment. Traces reporting addresses
are updated with the base address of each segment to preserve the "natural" address of
each register. In order to avoid many offset computations and degrade code readability,
also split the register array to support on array for each segment.

Update macros to use the select the proper segment while addressing the register content.

To address the register bitfield permutations/widening, as the scope of these changes is
limited, simply dispatch handling based on the selected version.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>