[ot] .github/workflows: build_test.yaml: Ignore LC on EG smoketest
In the smoketest there is no loaded OTP, so the LC state will be RAW.
This means it will not deliver a CPU_EN signal to Ibex to allow
instruction execution, which is needed for the smoketest.

Now that we do not enable this property by default, we should enable it
here in the smoketest.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>