[ot] hw/opentitan: ot_spi_device: Invert CSb status
The SPI Device's `STATUS.csb` field is the "Direct input of CSb signal".
The current logic reports a 1 if the CS is active (i.e. we are not idle
or in some error state, the SPI bus is active), but /CS is active low
and as such it should report a 0 when active, and a 1 when not.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>