[ot] hw/opentitan: ot_eg_pad_ring: Add initial stubbed Earlgrey Padring
Only contains the Power-on Reset (PoR) pad for now. While the PoR egress
IRQ is maintained, a separate PoR IRQ is introduced for linking with the
Rstmgr to allow to correctly specify the reset reason. This is hardcoded
with a separate dedicated signal for now to avoid overcomplicating the
interface, especially given that refactoring will likely be needed
anyhow for MIO pads for pinmux.

Also adds TODOs detailing the remaining steps to support the missing
features from Earlgrey's pad ring (in loose detail).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>