[ot] hw/opentitan: ot_spi_device: Maintain CSb across transactions
Rework handling of the emulated CSb (Chip Select) value to be performed
manually rather than using the current bus state.

The current behaviour assumes that CS is only active when the bus is not
idle or in an error state, but in the case where we get an incoming
transaction with a header with the `release` bit not set, this will lead
to the host thinking that the CS is held asserted between transactions
whereas the device thinks that it is released between them.

Move to manually handling this CSb state on bus transition to mimic the
previous behaviour with fixes for holding the CS across transactions.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>